+++ src/main/java/gtclassic/common/tile/GTTileDisassembler.java
@@ -1,69 +1,86 @@
 package gtclassic.common.tile;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collections;
+import java.util.Iterator;
+import java.util.LinkedHashSet;
+import java.util.LinkedList;
 import java.util.List;
+import java.util.Set;
 
 import gtclassic.GTMod;
 import gtclassic.api.helpers.GTHelperStack;
-import gtclassic.api.helpers.int3;
+import gtclassic.api.recipe.GTRecipeMachineHandler;
+import gtclassic.api.recipe.GTRecipeMultiInputList;
+import gtclassic.api.recipe.GTRecipeMultiInputList.MultiRecipe;
+import gtclassic.api.tile.GTTileBaseMachine;
 import gtclassic.common.container.GTContainerDisassembler;
 import gtclassic.common.gui.GTGuiMachine.GTDisassemblerGui;
-import ic2.api.classic.network.adv.NetworkField;
-import ic2.api.classic.tile.machine.IProgressMachine;
+import ic2.api.classic.item.IMachineUpgradeItem;
+import ic2.api.classic.item.IMachineUpgradeItem.UpgradeType;
+import ic2.api.classic.recipe.RecipeModifierHelpers.IRecipeModifier;
+import ic2.api.classic.recipe.machine.MachineOutput;
+import ic2.api.recipe.IRecipeInput;
 import ic2.core.RotationList;
-import ic2.core.block.base.tile.TileEntityElecMachine;
-import ic2.core.block.base.util.info.ProgressInfo;
-import ic2.core.inventory.base.IHasGui;
+import ic2.core.block.base.util.output.MultiSlotOutput;
 import ic2.core.inventory.container.ContainerIC2;
+import ic2.core.inventory.filters.ArrayFilter;
+import ic2.core.inventory.filters.BasicItemFilter;
 import ic2.core.inventory.filters.CommonFilters;
+import ic2.core.inventory.filters.IFilter;
+import ic2.core.inventory.filters.MachineFilter;
 import ic2.core.inventory.management.AccessRule;
 import ic2.core.inventory.management.InventoryHandler;
 import ic2.core.inventory.management.SlotType;
-import ic2.core.inventory.transport.IItemTransporter;
-import ic2.core.inventory.transport.TransporterManager;
+import ic2.core.item.recipe.entry.RecipeInputItemStack;
 import ic2.core.platform.registry.Ic2Items;
+import ic2.core.platform.registry.Ic2Sounds;
 import net.minecraft.client.gui.GuiScreen;
 import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.init.Items;
 import net.minecraft.init.SoundEvents;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.crafting.IRecipe;
-import net.minecraft.tileentity.TileEntity;
-import net.minecraft.util.ITickable;
+import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.SoundCategory;
 import net.minecraftforge.fml.common.registry.ForgeRegistries;
 
-public class GTTileDisassembler extends TileEntityElecMachine implements ITickable, IProgressMachine, IHasGui {
+public class GTTileDisassembler extends GTTileBaseMachine {
 
 	protected static final int[] slotInputs = { 0 };
 	protected static final int[] slotOutputs = { 1, 2, 3, 4, 5, 6, 7, 8, 9 };
+	int slotFuel = 10;
+	public IFilter filter = new MachineFilter(this);
 	public static final ResourceLocation GUI_LOCATION = new ResourceLocation(GTMod.MODID, "textures/gui/disassembler.png");
-	@NetworkField(index = 7)
-	float progress = 0;
+	public static final GTRecipeMultiInputList RECIPE_LIST = new GTRecipeMultiInputList("gt.disassembler", 16);
 
 	public GTTileDisassembler() {
-		super(11, 32);
-		maxEnergy = 2000;
-		addGuiFields("progress");
-		addInfos(new ProgressInfo(this));
+		super(11, 4, 16, 100, 32);
+		this.maxEnergy = 10000;
+		setFuelSlot(slotFuel);
 	}
 
 	@Override
 	protected void addSlots(InventoryHandler handler) {
 		handler.registerDefaultSideAccess(AccessRule.Both, RotationList.ALL);
+		handler.registerDefaultSlotAccess(AccessRule.Both, slotFuel);
 		handler.registerDefaultSlotAccess(AccessRule.Import, slotInputs);
 		handler.registerDefaultSlotAccess(AccessRule.Export, slotOutputs);
 		handler.registerDefaultSlotsForSide(RotationList.DOWN.invert(), slotInputs);
 		handler.registerDefaultSlotsForSide(RotationList.UP.invert(), slotOutputs);
-		handler.registerInputFilter(CommonFilters.Anything, slotInputs);
+		handler.registerInputFilter(filter, slotInputs);
+		handler.registerInputFilter(new ArrayFilter(CommonFilters.DischargeEU, new BasicItemFilter(Items.REDSTONE), new BasicItemFilter(Ic2Items.suBattery)), slotFuel);
+		handler.registerOutputFilter(CommonFilters.NotDischargeEU, slotFuel);
+		handler.registerSlotType(SlotType.Fuel, slotFuel);
 		handler.registerSlotType(SlotType.Input, slotInputs);
 		handler.registerSlotType(SlotType.Output, slotOutputs);
 	}
 
 	@Override
-	public boolean canInteractWith(EntityPlayer var1) {
-		return !this.isInvalid();
+	public Set<UpgradeType> getSupportedTypes() {
+		return new LinkedHashSet<>(Arrays.asList(UpgradeType.values()));
 	}
 
 	@Override
@@ -77,141 +94,146 @@ public class GTTileDisassembler extends TileEntityElecMachine implements ITickab
 	}
 
 	@Override
-	public boolean hasGui(EntityPlayer var1) {
-		return true;
+	public int[] getInputSlots() {
+		return slotInputs;
+	}
+
+	public ResourceLocation getGuiTexture() {
+		return GUI_LOCATION;
 	}
 
 	@Override
-	public void onGuiClosed(EntityPlayer var1) {
+	public IFilter[] getInputFilters(int[] slots) {
+		return new IFilter[] { filter };
 	}
 
 	@Override
-	public float getMaxProgress() {
-		return 100;
+	public boolean isRecipeSlot(int slot) {
+		return slot == 0;
 	}
 
 	@Override
-	public float getProgress() {
-		return this.progress;
+	public int[] getOutputSlots() {
+		return slotOutputs;
 	}
 
 	@Override
-	public void update() {
-		if (world.getTotalWorldTime() % 20 == 0) {
-			if (this.isActive) {
-				this.setActive(false);
+	public GTRecipeMultiInputList getRecipeList() {
+		return RECIPE_LIST;
+	}
+
+	@Override
+	public ResourceLocation getStartSoundFile() {
+		return Ic2Sounds.maceratorOp;
+	}
+
+	@Override
+	public void process(MultiRecipe recipe) {
+		MachineOutput output = recipe.getOutputs().copy();
+		for (ItemStack stack : output.getRecipeOutput(getWorld().rand, getTileData())) {
+			if (world.rand.nextFloat() > .2) {
+				outputs.add(new MultiSlotOutput(stack, getOutputSlots()));
+				onRecipeComplete();
 			}
-			this.progress = 0;
-			this.getNetwork().updateTileGuiField(this, "progress");
-			tryImportItems();
-			tryExportItems();
-			ItemStack slot0 = this.getStackInSlot(0);
-			if (slot0.isEmpty()) {
-				return;
-			}
-			if (slot0.isItemDamaged()) {
-				return;
-			}
-			for (int x = 1; x < 10; ++x) {
-				ItemStack stack = inventory.get(x);
-				if (!stack.isEmpty()) {
-					return;
-				}
-			}
-			for (IRecipe recipe : ForgeRegistries.RECIPES) {
-				ItemStack input = recipe.getRecipeOutput().copy();
-				int amount = recipe.getRecipeOutput().getCount();
-				if (GTHelperStack.isEqual(input, slot0) && slot0.getCount() >= amount) {
-					this.setActive(true);
-					this.progress = this.getMaxProgress();
-					this.getNetwork().updateTileGuiField(this, "progress");
-					List<ItemStack> outputList = new ArrayList<>();
-					for (int i = 0; i < recipe.getIngredients().size(); ++i) {
-						List<ItemStack> tempList = new ArrayList<>();
-						Collections.addAll(tempList, recipe.getIngredients().get(i).getMatchingStacks());
-						if (!tempList.isEmpty()) {
-							outputList.add(tempList.get(0).copy());
+		}
+		NBTTagCompound nbt = recipe.getOutputs().getMetadata();
+		boolean shiftContainers = nbt == null ? false : nbt.getBoolean(MOVE_CONTAINER_TAG);
+		List<ItemStack> inputs = getInputs();
+		List<IRecipeInput> recipeKeys = new LinkedList<IRecipeInput>(recipe.getInputs());
+		for (Iterator<IRecipeInput> keyIter = recipeKeys.iterator(); keyIter.hasNext();) {
+			IRecipeInput key = keyIter.next();
+			int count = key.getAmount();
+			for (Iterator<ItemStack> inputIter = inputs.iterator(); inputIter.hasNext();) {
+				ItemStack input = inputIter.next();
+				if (key.matches(input)) {
+					if (input.getCount() >= count) {
+						if (input.getItem().hasContainerItem(input)) {
+							if (!shiftContainers) {
+								continue;
+							}
+							ItemStack container = input.getItem().getContainerItem(input);
+							if (!container.isEmpty()) {
+								container.setCount(count);
+								outputs.add(new MultiSlotOutput(container, getOutputSlots()));
+							}
+						}
+						input.shrink(count);
+						count = 0;
+						if (input.isEmpty()) {
+							inputIter.remove();
+						}
+						keyIter.remove();
+						break;
+					}
+					if (input.getItem().hasContainerItem(input)) {
+						if (!shiftContainers) {
+							continue;
+						}
+						ItemStack container = input.getItem().getContainerItem(input);
+						if (!container.isEmpty()) {
+							container.setCount(input.getCount());
+							outputs.add(new MultiSlotOutput(container, getOutputSlots()));
 						}
 					}
-					// i think the recipe ingredients being a nonnulllist breaks this
-					// if (outputList.size() < 2) {
-					// return;
-					// }
-					for (int j = 0; j < outputList.size(); ++j) {
-						ItemStack outputStack = outputList.get(j).copy();
-						if (canItemBeReturned(outputStack) && world.rand.nextFloat() > .20F) {
-							this.setStackInSlot(j + 1, outputStack);
-						}
-					}
-					slot0.shrink(amount);
-					world.playSound((EntityPlayer) null, pos, SoundEvents.ENTITY_ITEM_BREAK, SoundCategory.BLOCKS, 0.5F, 1.0F);
-					return;
+					count -= input.getCount();
+					input.setCount(0);
+					inputIter.remove();
 				}
 			}
 		}
+		addToInventory();
+		if (supportsUpgrades) {
+			for (int i = 0; i < upgradeSlots; i++) {
+				ItemStack item = inventory.get(i + inventory.size() - upgradeSlots);
+				if (item.getItem() instanceof IMachineUpgradeItem) {
+					((IMachineUpgradeItem) item.getItem()).onProcessFinished(item, this);
+				}
+			}
+		}
+		shouldCheckRecipe = true;
+	}
+
+	@Override
+	public void onRecipeComplete() {
+		world.playSound((EntityPlayer) null, pos, SoundEvents.ENTITY_ITEM_BREAK, SoundCategory.BLOCKS, 0.05F, 1.0F);
+	}
+
+	public static void init() {
+		for (IRecipe recipe : ForgeRegistries.RECIPES) {
+			ItemStack input = recipe.getRecipeOutput().copy();
+			List<ItemStack> outputList = new ArrayList<>();
+			for (int i = 0; i < recipe.getIngredients().size(); ++i) {
+				List<ItemStack> tempList = new ArrayList<>();
+				Collections.addAll(tempList, recipe.getIngredients().get(i).getMatchingStacks());
+				if (!tempList.isEmpty()) {
+					ItemStack tempStack = tempList.get(0).copy();
+					if (canItemBeReturned(tempStack)) {
+						outputList.add(tempStack);
+					}
+				}
+			}
+			if (canInputBeUsed(input) && !outputList.isEmpty()) {
+				ItemStack[] arr = outputList.toArray(new ItemStack[0]);
+				addRecipe(new IRecipeInput[] { new RecipeInputItemStack(input) }, totalEu(5000), arr);
+			}
+		}
+	}
+
+	public static IRecipeModifier[] totalEu(int total) {
+		return GTRecipeMachineHandler.totalEu(RECIPE_LIST, total);
+	}
+
+	public static void addRecipe(IRecipeInput[] inputs, IRecipeModifier[] modifiers, ItemStack... outputs) {
+		GTRecipeMachineHandler.addRecipe(RECIPE_LIST, inputs, modifiers, outputs);
+	}
+
+	public static boolean canInputBeUsed(ItemStack input) {
+		return !input.isEmpty() && input.getCount() > 0 && !input.isItemDamaged()
+				&& !input.getItem().hasContainerItem(input);
 	}
 
 	public static boolean canItemBeReturned(ItemStack stack) {
 		return !GTHelperStack.isEqual(stack, Ic2Items.uuMatter.copy()) && !stack.isItemStackDamageable()
 				&& !stack.getUnlocalizedName().contains("bucket");
-	}
-
-	public TileEntity getImportTile() {
-		int3 dir = new int3(getPos(), getFacing());
-		return world.getTileEntity(dir.back(1).asBlockPos());
-	}
-
-	public TileEntity getExportTile() {
-		int3 dir = new int3(getPos(), getFacing());
-		return world.getTileEntity(dir.forward(1).asBlockPos());
-	}
-
-	public void tryImportItems() {
-		IItemTransporter slave = TransporterManager.manager.getTransporter(getImportTile(), true);
-		if (slave == null) {
-			return;
-		}
-		IItemTransporter controller = TransporterManager.manager.getTransporter(this, true);
-		int limit = 64;
-		for (int i = 0; i < limit; ++i) {
-			ItemStack stack = slave.removeItem(CommonFilters.Anything, getFacing(), 1, false);
-			if (stack.isEmpty()) {
-				break;
-			}
-			ItemStack added = controller.addItem(stack, getFacing().getOpposite(), true);
-			if (added.getCount() <= 0) {
-				break;
-			}
-			slave.removeItem(CommonFilters.Anything, getFacing(), 1, true);
-		}
-	}
-
-	public void tryExportItems() {
-		IItemTransporter slave = TransporterManager.manager.getTransporter(getExportTile(), true);
-		if (slave == null) {
-			return;
-		}
-		IItemTransporter controller = TransporterManager.manager.getTransporter(this, true);
-		int limit = 64;
-		for (int i = 0; i < limit; ++i) {
-			ItemStack stack = controller.removeItem(CommonFilters.Anything, getFacing(), 1, false);
-			if (stack.isEmpty()) {
-				break;
-			}
-			ItemStack added = slave.addItem(stack, getFacing().getOpposite(), true);
-			if (added.getCount() <= 0) {
-				break;
-			}
-			controller.removeItem(CommonFilters.Anything, getFacing(), 1, true);
-		}
-	}
-
-	@Override
-	public boolean supportsNotify() {
-		return true;
-	}
-
-	public ResourceLocation getGuiTexture() {
-		return GUI_LOCATION;
 	}
 }