+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # 1.07 (Not Released)
 - Completely refactored the entire project to now have a dedicated API although this doesnt change the mod for players at all.
 - Changed JEI system so addons can handle JEI automatically if they want, this wont effect gameplay at all
+- Fixed JEI now displays fluid inputs as fluids not fluid containers.
+- Added a Disassembler machine that gives an 80% return on each item uncrafting something, uses 5000EU per operation.
 - Added Phosphorus dust to the base mod, its a byproduct of Netherrack by default.
 - Added Sulfur dust to the base mod, its a byproduct of Netherrack by default.
 - Added Sulfuric Acid to the base mod, no use for it atm.